+++ src/FSharp.CosmosDb/Types.fs
@@ -1,6 +1,54 @@
 namespace FSharp.CosmosDb
 
-open Azure.Cosmos
+open Microsoft.Azure.Cosmos
+open System.Threading
+open System.Threading.Tasks
+open System.Collections.Concurrent
+open System
+open System.Collections.Generic
+
+module internal Caching =
+    let private clientCache =
+        ConcurrentDictionary<string, CosmosClient>()
+
+    let inline private tryGetOption key (cd: ConcurrentDictionary<_, _>) =
+        if cd.ContainsKey key then
+            Some(cd.[key])
+        else
+            None
+
+    let fromConnStr connStr clientOps =
+        maybe {
+            let! cs = connStr
+
+            return
+                clientCache
+                |> tryGetOption cs
+                |> Option.defaultWith
+                    (fun () ->
+                        let client = new CosmosClient(cs, clientOps)
+                        clientCache.[cs] <- client
+                        client)
+        }
+
+    let fromKey host' accessKey' clientOps =
+        maybe {
+            let! host = host'
+            let! accessKey = accessKey'
+
+            let connStr = sprintf "%s%s" host accessKey
+
+            return
+                clientCache
+                |> tryGetOption connStr
+                |> Option.defaultWith
+                    (fun () ->
+                        let client =
+                            new CosmosClient(host, accessKey, clientOps)
+
+                        clientCache.[connStr] <- client
+                        client)
+        }
 
 type ConnectionOperation =
     { Options: CosmosClientOptions option
@@ -10,6 +58,27 @@ type ConnectionOperation =
       ConnectionString: string option
       DatabaseId: string option
       ContainerName: string option }
+
+    member internal this.GetClient() =
+        let clientOps =
+            match this.Options with
+            | Some op -> op
+            | None -> CosmosClientOptions()
+
+        let client =
+            if this.FromConnectionString then
+                Caching.fromConnStr this.ConnectionString clientOps
+            else
+                Caching.fromKey this.Endpoint this.AccessKey clientOps
+
+        match client with
+        | Some client -> client
+        | None -> failwith "No connection information provided"
+
+    interface IDisposable with
+        override this.Dispose() =
+            let client = this.GetClient()
+            client.Dispose()
 
 // --- Operation Types --- //
 type QueryOp<'T> =
@@ -53,3 +122,13 @@ type ContainerOperation<'T> =
     | Upsert of UpsertOp<'T>
     | Read of ReadOp<'T>
     | Replace of ReplaceOp<'T>
+
+type ChangeFeedOptions<'T> =
+    { Connection: ConnectionOperation
+      Processor: string
+      OnChange: IReadOnlyCollection<'T> -> CancellationToken -> Task
+      InstanceName: string option
+      LeaseContainer: ConnectionOperation option
+      PollingInterval: TimeSpan option
+      StartTime: DateTime option
+      MaxItems: int option }